Moves observation wrapping to controller, fixes hal & json rendering.
